gwproxy: Initial epoll code split
Initial work to split epoll specific code into a separate file. It's
also the first step in adding io_uring support.
