+++ src/gwproxy/gwproxy.c
@@ -7,6 +7,7 @@
 #include <gwproxy/gwproxy.h>
 #include <gwproxy/common.h>
 #include <gwproxy/log.h>
+#include <gwproxy/ev/epoll.h>
 
 #include <stdio.h>
 #include <stdlib.h>
@@ -490,96 +491,14 @@ static void gwp_ctx_free_thread_sock(struct gwp_wrk *w)
 	}
 }
 
-__cold
-static int gwp_ctx_init_thread_epoll(struct gwp_wrk *w)
+static int gwp_ctx_init_thread_event(struct gwp_wrk *w)
 {
-	struct epoll_event ev, *events;
-	struct gwp_ctx *ctx = w->ctx;
-	int ep_fd, ev_fd, r;
-
-	ep_fd = __sys_epoll_create1(EPOLL_CLOEXEC);
-	if (ep_fd < 0) {
-		r = ep_fd;
-		pr_err(&w->ctx->lh, "Failed to create epoll instance: %s\n",
-			strerror(-r));
-		return r;
-	}
-
-	ev_fd = __sys_eventfd(0, EFD_CLOEXEC | EFD_NONBLOCK);
-	if (ev_fd < 0) {
-		r = ev_fd;
-		pr_err(&w->ctx->lh, "Failed to create eventfd: %s\n", strerror(-r));
-		goto out_close_ep_fd;
-	}
-
-	w->evsz = 512;
-	events = calloc(w->evsz, sizeof(*events));
-	if (!events) {
-		r = -ENOMEM;
-		pr_err(&w->ctx->lh, "Failed to allocate memory for events: %s\n",
-			strerror(-r));
-		goto out_close_ev_fd;
-	}
-
-	w->ev_fd = ev_fd;
-	w->ep_fd = ep_fd;
-	w->events = events;
-
-	memset(&ev, 0, sizeof(ev));
-	ev.events = EPOLLIN;
-	ev.data.u64 = EV_BIT_EVENTFD;
-	r = __sys_epoll_ctl(ep_fd, EPOLL_CTL_ADD, ev_fd, &ev);
-	if (unlikely(r))
-		goto out_free_events;
-
-	ev.events = EPOLLIN;
-	ev.data.u64 = EV_BIT_ACCEPT;
-	r = __sys_epoll_ctl(ep_fd, EPOLL_CTL_ADD, w->tcp_fd, &ev);
-	if (unlikely(r))
-		goto out_free_events;
-
-	if (w->idx == 0 && (ctx->ino_fd >= 0)) {
-		ev.events = EPOLLIN;
-		ev.data.u64 = EV_BIT_SOCKS5_AUTH_FILE;
-		r = __sys_epoll_ctl(ep_fd, EPOLL_CTL_ADD, ctx->ino_fd, &ev);
-		if (unlikely(r))
-			goto out_free_events;
-	}
-
-	pr_dbg(&w->ctx->lh, "Worker %u epoll (ep_fd=%d, ev_fd=%d)", w->idx,
-		ep_fd, ev_fd);
-	return 0;
-
-out_free_events:
-	free(events);
-	w->events = NULL;
-out_close_ev_fd:
-	__sys_close(ev_fd);
-out_close_ep_fd:
-	__sys_close(ep_fd);
-	w->ev_fd = w->ep_fd = -1;
-	return r;
+	return gwp_ctx_init_thread_epoll(w);
 }
 
-__cold
-static void gwp_ctx_free_thread_epoll(struct gwp_wrk *w)
+static void gwp_ctx_free_thread_event(struct gwp_wrk *w)
 {
-	if (w->ev_fd >= 0) {
-		__sys_close(w->ev_fd);
-		pr_dbg(&w->ctx->lh, "Worker %u eventfd closed (fd=%d)", w->idx,
-		       w->ev_fd);
-		w->ev_fd = -1;
-	}
-
-	if (w->ep_fd >= 0) {
-		__sys_close(w->ep_fd);
-		pr_dbg(&w->ctx->lh, "Worker %u epoll closed (fd=%d)", w->idx,
-		       w->ep_fd);
-		w->ep_fd = -1;
-	}
-
-	free(w->events);
-	w->events = NULL;
+	gwp_ctx_free_thread_epoll(w);
 }
 
 __cold
@@ -595,9 +514,9 @@ static int gwp_ctx_init_thread(struct gwp_wrk *w,
 		return r;
 	}
 
-	r = gwp_ctx_init_thread_epoll(w);
+	r = gwp_ctx_init_thread_event(w);
 	if (r < 0) {
-		pr_err(&w->ctx->lh, "Failed to initialize epoll for worker %u: %s\n",
+		pr_err(&w->ctx->lh, "Failed to initialize event for worker %u: %s\n",
 			w->idx, strerror(-r));
 		goto out_free_sock;
 	}
@@ -658,7 +577,7 @@ static void gwp_ctx_free_thread(struct gwp_wrk *w)
 	if (w->idx > 0)
 		pthread_join(w->thread, NULL);
 	gwp_ctx_free_thread_sock_pairs(w);
-	gwp_ctx_free_thread_epoll(w);
+	gwp_ctx_free_thread_event(w);
 	gwp_ctx_free_thread_sock(w);
 }
